+++ CHANGES.rst
@@ -22,6 +22,7 @@ Changelog
 - the `classes_to_str` and `get_class_name` methods can remove `builtins.` via the clean=True parameter now
 - introduced the `InifiniteReader` mixin which automatically disables batch mode
 - added support for caching plugins managed by the `ClassRegistry` via the `ClassCache` class (one for each class hierarchy)
+- the `execute` method now supports custom pre-initialization and post-finalization method hooks
 
 
 0.2.21 (2025-08-19)